+++ lib/models/subscription.dart
@@ -1,0 +1,10 @@
+class Subscription {
+  String charity;
+  String charityId;
+  int goals;
+  String name;
+  String teamName;
+  String time;
+
+  Subscription(this.charity, this.charityId, this.goals, this.name, this.teamName, this.time);
+}