+++ CHANGELOG.md
@@ -141,4 +141,4 @@ __15/08/18__
 
 - Profile Page has been started, although currently the form collapses whenever the keyboard is opened
 
--
+- Added Subscription model and set up navigation to and from Subscription Details Page from Subscriptions